Transform `__proto__` property access
Transform `__proto__` property access to the equivalent expression using
`Object.getPrototypeOf`.

     var proto1 = [].__proto__;
     var proto2 = []['__proto__'];

becomes:

     var proto1 = Object.getPrototypeOf([]);
     var proto2 = Object.getPrototypeOf([]);

Motivation: the `__proto__` property is an AnnexB extension, so tests
should not rely on its behavior unless it is the explicit subject of the
tests. The V8 test suite uses the property extensively, making an
automated transformation worthwhile.